Add login_required decorators to pages that require a logged in user. Add redirect to private views to prevent non-superusers from accessing those views

+++ products/views.py
@@ -3,6 +3,7 @@ from django.db.models import Q
 from .models import Product, ProductSizesStock
 from .forms import ProductForm
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 
 def all_products(request):
@@ -96,7 +97,12 @@ def product_details(request, product_id):
     return render(request, 'products/product_details.html', context)
 
 
+@login_required
 def add_product(request):
+
+    if not request.user.is_superuser:
+        messages.error(request, 'sorry this page is private')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -118,7 +124,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
+
+    if not request.user.is_superuser:
+        messages.error(request, 'sorry this page is private')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == 'POST':
@@ -143,7 +155,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
+
+    if not request.user.is_superuser:
+        messages.error(request, 'sorry this page is private')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted')
